+++ thu.sh
@@ -754,7 +754,8 @@ thumb_create_from_video () {
     vid_wh_native=$(video_resolution_ffmpeg_parse "$vid_ffmpeg_output")
     vid_wh_scaled=$(wh_scaled_get "$vid_wh_native" "$vid_wh_max")
     vid_frame_ss=$(($vid_duration_ss / 5))
-    vid_thumb_path=$(cachedir_path_get "$cachedir" "video" "w h" ".png")
+    vid_thumb_path=$(
+        cachedir_path_get "$cachedir" "video" "$vid_wh_scaled" ".png")
 
     if [[ -z "$is_cmd_ffmpeg" ]]; then
         fail "$msg_cmd_not_found_ffmpeg";